pyproject.toml
[project]
name = "kepegawaian-cronjob"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.117.1",
    "pandas>=2.3.2",
    "pymysql-pool>=0.5.0",
    "python-dotenv>=1.1.1",
]

[dependency-groups]
dev = [
    "icecream>=2.1.8",
    "openpyxl>=3.1.5",
    "pandas-stubs>=2.3.2.250827",
]
